The DNS module was returning a list containing "undef" on resolve failure, but the callback in ServerOut seems to be expecting a list of results that is empty if resolution fails.
Switched the DNS module to return an empty list in the fail case, since this
seems to me to be the most sensible approach for a function that can potentially
return multiple results. Should just evaluate to undef in the scalar case anyway.

Also, was an apparently incorrect callback "fail" in the failure case, which I've
switched to "error". Everything now appears to be hunky-dory, though I must admit
I'm not 100% confident that this is correct as there is still lots of code I've
not even looked at...



git-svn-id: http://code.sixapart.com/svn/djabberd/trunk@290 43dd9337-660f-0410-bd32-e7601923a1a1