Basic support for in-process "components".
Here I'm defining a component as an agent that handles JIDs for an entire domain. This is differs from a bot, which handles only a single node.

A plugin is provided for making a component appear as a subdomain of a vhost. However, the component subclasses don't really care what calls them so you could theoretically make a Delivery plugin that turns an entire VHost into a component, or whatever.


git-svn-id: http://code.sixapart.com/svn/djabberd/trunk@550 43dd9337-660f-0410-bd32-e7601923a1a1