Add DJabberd::Socket, a cut down copy of Danga::Socket::AnyEvent.
This replaces use of Danga::Socket in DJabberd::Connection, preserving
the expected interface but using AnyEvent as the event loop abstraction.

This should be cut down even more as the rest of the codebase is changed
to use AnyEvent facilities directly. Certainly all of the methods that
aren't related to instances should eventually be removed.