Restructure debian packaging.
This commit restructures the debian packaging. First it splits off
the modules from the djabberd package which allows installing all
required libraries w/o getting another deamon (e.g. as build
dependency for building other modules or custom jabberds). Second
it adds the missing wiring to the djabberd to get an jabber
server which is easy to set up. Including propper logging and
an ssl certificate.