Add c2s IPv6 support
This commit adds IPv6 c2s (listening) support to the server using
IO::Socket::INET6 instead of IO::Socket::INET.

It also relaxes the IP(v4) check of DJabberd::Util::as_bind_addr
and introduces some new testcases.

This patch works for me, but has the drawback of requiring IPv6
capabilities even on hosts w/o (a need for) IPv6.

As IPv4 is facing its end-of-life I deem this is OK, but YMMV.