Add IPv6 support
This commit adds IPv6 support to the server by using IO::Socket::INET6
instead of IO::Socket::INET and by relaxing the IP check in
DJabberd::Util::as_bind_addr.

This patch works for me, but has the drawback of requiring IPv6
capabilities even on hosts w/o (a need for) IPv6.

As IPv4 is facing its end-of-life I deem this is OK, but YMMV.